WIP: Refactor types
