+++ hkmc2/shared/src/main/scala/hkmc2/bbml/bbML.scala
@@ -10,138 +10,6 @@ import semantics.*, semantics.Term.*
 import syntax.*
 import Tree.*
 
-sealed abstract class TypeArg:
-  lazy val lvl: Int
-  def map(f: Type => Type): TypeArg
-
-sealed trait GeneralType:
-  lazy val isPoly: Bool
-  lazy val lvl: Int
-  def toString(): String
-  def monoOr(fallback: => Type): Type
-
-case class Wildcard(in: Type, out: Type) extends TypeArg {
-  private def printWhenSet(t: Type, in: Bool) = t match
-    case Type.Top if !in => N
-    case Type.Bot if in => N
-    case _ => S(if in then s"in $t" else s"out $t")
-  def map(f: Type => Type): Wildcard = Wildcard(f(in), f(out))
-
-  override def toString(): String = s"in $in out $out"
-  override lazy val lvl: Int = in.lvl.max(out.lvl)
-}
-
-object Wildcard:
-  def in(ty: Type): Wildcard = Wildcard(ty, Type.Top)
-  def out(ty: Type): Wildcard = Wildcard(Type.Bot, ty)
-  def empty: Wildcard = Wildcard(Type.Bot, Type.Top)
-
-enum Type extends TypeArg with GeneralType:
-  case ClassType(name: ClassSymbol, targs: Ls[TypeArg])
-  case InfVar(vlvl: Int, uid: Uid[InfVar], state: VarState)
-  case FunType(args: Ls[Type], ret: Type, eff: Type)
-  case ComposedType(lhs: Type, rhs: Type, pol: Bool) // * positive -> union
-  case NegType(ty: Type)
-  case Top
-  case Bot
-  override def toString(): String = this match
-    case ClassType(name, targs) =>
-      if targs.isEmpty then s"${name.nme}" else s"${name.nme}[${targs.mkString(", ")}]"
-    case InfVar(lvl, uid, _) => s"α${uid}_$lvl"
-    case FunType(args, ret, eff) => s"(${args.mkString(", ")}) ->{${eff}} ${ret}"
-    case ComposedType(lhs, rhs, pol) => s"(${lhs}) ${if pol then "∨" else "∧"} (${rhs})"
-    case NegType(ty) => s"¬(${ty})"
-    case Top => "⊤"
-    case Bot => "⊥"
-  lazy val lvl: Int = this match
-    case ClassType(name, targs) =>
-      targs.map(_.lvl).maxOption.getOrElse(0)
-    case InfVar(lvl, _, _) => lvl
-    case FunType(args, ret, eff) =>
-      (ret :: eff :: args).map(_.lvl).max
-    case ComposedType(lhs, rhs, _) =>
-      lhs.lvl.max(rhs.lvl)
-    case NegType(ty) => ty.lvl
-    case Top | Bot => 0
-  lazy val isPoly: Bool = false
-  def & (that: Type): Type = this match
-    case Top => that
-    case Bot => Bot
-    case _ => that match
-      case Top => this
-      case Bot => Bot
-      case ComposedType(l, r, false) => this & l & r
-      case _ => if this === that
-        then this
-        else ComposedType(this, that, false)
-  def | (that: Type): Type = this match
-    case Top => Top
-    case Bot => that
-    case _ => that match
-      case Top => Top
-      case Bot => this
-      case ComposedType(l, r, true) => this | l | r
-      case _ => if this === that
-        then this
-        else ComposedType(this, that, true)
-  def ! : Type = this match
-    case Top => Bot
-    case Bot => Top
-    case NegType(res) => res
-    case ComposedType(l, r, true) => l.! & r.!
-    case ComposedType(l, r, false) => l.! | r.!
-    case _ => NegType(this)
-  def isSmall: Bool = this match
-    // (copilot:)
-    // case ClassType(_, targs) => targs.forall(_.isSmall)
-    // case InfVar(_, _, _) => true
-    // case FunType(args, ret, eff) => args.forall(_.isSmall) && ret.isSmall && eff.isSmall
-    // case ComposedType(lhs, rhs, _) => lhs.isSmall && rhs.isSmall
-    // case NegType(ty) => ty.isSmall
-    // case PolymorphicType(tv, body) => tv.forall(_.isSmall) && body.isSmall
-    // case Top | Bot => true
-    case _ => true
-  def map(f: Type => Type): Type = this match
-    case ClassType(name, targs) => ClassType(name, targs.map(_.map(f)))
-    case InfVar(lvl, uid, state) => f(this)
-    case FunType(args, ret, eff) => FunType(args.map(f), f(ret), f(eff))
-    case ComposedType(lhs, rhs, pol) => Type.mkComposedType(f(lhs), f(rhs), pol)
-    case NegType(ty) => Type.mkNegType(f(ty))
-    case Top | Bot => f(this)
-  def monoOr(fallback: => Type): Type = this
-object Type:
-  def mkComposedType(lhs: Type, rhs: Type, pol: Bool): Type = if pol
-    then lhs | rhs
-    else lhs & rhs
-  def mkNegType(ty: Type): Type = ty.!
-
-// TODO: bounds
-case class PolyType(tv: Ls[Type.InfVar], body: GeneralType) extends GeneralType:
-  lazy val isPoly: Bool = true
-  lazy val lvl: Int = (body :: tv).map(_.lvl).max
-  override def toString(): String = s"forall ${tv.mkString(", ")}: $body"
-  def monoOr(fallback: => Type): Type = fallback
-  def map(f: GeneralType => GeneralType): GeneralType = ???
-
-case class PolyFunType(args: Ls[GeneralType], ret: GeneralType, eff: Type) extends GeneralType:
-  lazy val isPoly: Bool = (ret :: args).exists(_.isPoly)
-  lazy val lvl: Int = (ret :: eff :: args).map(_.lvl).max
-  override def toString(): String = s"(${args.mkString(", ")}) ->{${eff}} ${ret}"
-  lazy val mono: Option[Type.FunType] = if isPoly then N else
-    Some(Type.FunType(args.map {
-      case t: Type => t
-      case pf: PolyFunType => pf.mono.get
-    }, ret match {
-      case t: Type => t
-      case pf: PolyFunType => pf.mono.get
-    }, eff))
-  def monoOr(fallback: => Type): Type = mono.getOrElse(fallback)
-  def map(f: GeneralType => GeneralType): GeneralType = ???
-
-class VarState:
-  var lowerBounds: Ls[Type] = Nil
-  var upperBounds: Ls[Type] = Nil
-
 object InfVarUid extends Uid.Handler[Type.InfVar]
 
 type EnvType = HashMap[Uid[Symbol], GeneralType]
